@@ -1,15 +1,14 @@
 import React from 'react';
-
 import BlessPointsCell from '../BlessPointsCell';
-
 import styles from './BlessEffectsRows.module.scss';
 
 function BlessEffectsRows(props) {
-    const { effects } = props;
+    const { effects, onBlessSelect, isExternalWindow } = props;
     const rows = effects.map(effect => {
-        const { id, scales, name, shortDescription, incarnate } = effect;
+        const { id, scales, name, shortDescription, incarnate, multi } = effect;
+
 	return (
-	    <tr key={id} className={styles.row}>
+	    <tr key={id} className={styles.row} onClick={() => onBlessSelect?.(effect)}>
               <BlessPointsCell effect={effect}/>
               <td className={styles.scales}>
 		{
@@ -21,9 +20,14 @@ function BlessEffectsRows(props) {
 			+ (scales?.magic        ? (`Magic: ${       scales.magic        } `) : "")
 		}
               </td>
-              <td className={styles.name}>{name}</td>
+              <td className={styles.name}>{name} {multi && " (*)"}</td>
               <td className={styles.description}>{shortDescription}</td>
               <td className={styles.incarnate}>{incarnate ? "yes" : ""}</td>
+              {!isExternalWindow && (
+                <td className={styles.add_button}>
+                  <p>Add</p>
+                </td>
+              )}
 	    </tr>
 	);
     });
