@@ -5,14 +5,13 @@ import {filterPretendersByImprisonment} from './filterPretendersByImprisonment';
 import {filterPretendersByChassis} from './filterPretendersByChassis';
 import { blessOptimizer } from './blessOptimizer';
 import {pretenderCost} from './pretenderCost';
-import { sum } from 'ramda';
-
+import Export from './Export';
 import styles from './Pretenders.module.scss';
 
 function Pretenders(props) {
 
   const {
-      f, a, w, e, s, d, n, g, b,  
+      f, a, w, e, s, d, n, g, b,
       nations,
       nationId,
       pretenders,
@@ -26,14 +25,17 @@ function Pretenders(props) {
       blessBonus,
       openPretenderOptimizer,
       closePretenderOptimizer,
-      isPretenderOptimizerOpen
+      isPretenderOptimizerOpen,
+      selectedBlesses
   } = props;
 
   const paths = {f, a, w, e, s, d, n, g, b};
   
   const points = 450;
 
-  const totalCostOfScales = sum(Object.values(scalesCosts));
+  const totalCostOfScales = Object.values(scalesCosts).reduce(
+    (acc, curr) => acc + curr, 0
+  );
   const pointsLeftWithoutPretenders = points + pointsForImprisonment - totalCostOfScales;
 
   const filteredPretenderIds = filterPretendersByChassis(
@@ -51,7 +53,7 @@ function Pretenders(props) {
   if (isPretenderOptimizerOpen) {
     const filteredPretenderIdsByCost = filteredPretenderIds.filter(pretenderId => {
       return pointsLeftWithoutPretenders - pretenderCost(pretenders[pretenderId], nations[nationId], paths, dominion) >= 0;
-    })
+    });
     
     const pretenderRows = filteredPretenderIdsByCost.map(pretenderId => {
       const pretender = pretenders[pretenderId];
@@ -251,6 +253,17 @@ function Pretenders(props) {
           <td className={styles[`table_cell${data.n > 0 ? "_n" : ""}`]}>{data.n > 0 ? `N${data.n}` : ""}</td>
           <td className={styles[`table_cell${data.g > 0 ? "_g" : ""}`]}>{data.g > 0 ? `G${data.g}` : ""}</td>
           <td className={styles[`table_cell${data.b > 0 ? "_b" : ""}`]}>{data.b > 0 ? `B${data.b}` : ""}</td>
+          {nationId !== 0 && nationId !== "0" && (
+            <Export
+              nationId={nationId}
+              pretender={data}
+              paths={paths}
+              dominion={dominion}
+              imprisonment={imprisonment}
+              scales={scales}
+              selectedBlesses={selectedBlesses}
+            />
+          )}
         </tr>
     );
       });
@@ -268,6 +281,9 @@ function Pretenders(props) {
           <th className={styles.table_header}>Dominion</th>
           <th className={styles.table_header}>Scales</th>
           <th className={styles.table_header} colSpan="9">Magic</th>
+          {nationId !== 0 && nationId !== "0" && (
+            <th className={styles.table_header}>Export</th>
+          )}
         </tr>
       </thead>
       <tbody id="pretenders-table__body">
