@@ -1,17 +1,13 @@
 import getNations from '../getNations';
-import { defaultTo } from 'ramda';
 
 const getScaleLimits = (nationId) => {
-    const scalelimits = defaultTo(
-        {},
-        getNations()[nationId].scalelimits
-    );
-    const order = defaultTo(0, scalelimits.order);
-    const productivity = defaultTo(0, scalelimits.productivity);
-    const heat = defaultTo(0, scalelimits.heat);
-    const growth = defaultTo(0, scalelimits.growth);
-    const fortune = defaultTo(0, scalelimits.fortune);
-    const magic = defaultTo(0, scalelimits.magic);
+  const scalelimits = getNations()[nationId]?.scalelimits ?? {};
+  const order         = scalelimits.order         ?? 0;
+  const productivity  = scalelimits.productivity  ?? 0;
+  const heat          = scalelimits.heat          ?? 0;
+  const growth        = scalelimits.growth        ?? 0;
+  const fortune       = scalelimits.fortune       ?? 0;
+  const magic         = scalelimits.magic         ?? 0;
     return {
         order, productivity, heat,
         growth, fortune, magic
